requirements.txt
langchain==0.2.15
langgraph==0.2.19
langchain-openai==0.1.8
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.22
google-auth==2.29.0
google-api-python-client==2.122.0
pydantic==2.7.0
httpx==0.27.0